requirements.txt
altgraph==0.17.4
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
eventlet==0.40.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.5.1
fonttools==4.54.1
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==3.3.0
kiwisolver==1.4.7
macholib==1.16.3
MarkupSafe==3.0.1
matplotlib==3.9.2
networkx==3.4.1
numpy==2.1.2
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
pluggy==1.5.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.9.0
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-socketio==5.13.0
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.2.0
simple-websocket==1.1.0
six==1.16.0
tk==0.1.0
trueskill==0.4.5
urllib3==2.2.3
Werkzeug==3.0.6
wsproto==1.2.0
